Conflicts:
	OpenGlTest2/OculusApp.cpp
	OpenGlTest2/OculusApp.h
	OpenGlTest2/OpenGlTest2.vcxproj
	OpenGlTest2/OpenGlTest2.vcxproj.filters

+++ OpenGlTest2/Vector2.h
@@ -14,8 +14,8 @@
 
 typedef struct
 {
-    GLfloat x;
-	GLfloat y;
+    GLdouble x;
+	GLdouble y;
     
 }Vector2;
